% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_keyword.R
\name{extract_keyword}
\alias{extract_keyword}
\title{Extract text containing a keyword from a PDF}
\usage{
extract_keyword(file, keyword, nbefore = 0L, nafter = 0L)
}
\arguments{
\item{file}{A character vector of the PDF file names to search.}

\item{keyword}{The keyword to look for in the PDF(s).}

\item{nbefore}{The number of words to include before the keyword. The
default is 0.}

\item{nafter}{The number of words to include after the keyword. The default
is 0.}
}
\value{
A data frame with three variables: \code{PDF_file}, \code{PDF_page},
  and \code{Keyword_extract}.
}
\description{
\code{extract_keyword} finds all occurrences of a given \code{keyword} in
the text of one or more PDFs, and returns a data frame containing a row for
each occurrence found. Each row of the data frame is made up of a string
containing the keyword in its immediate context, the page number of the PDF
on which the keyword was found, and the file name of the PDF.
}
\examples{
\dontrun{
extract_keyword("example.pdf", "foo")
}
}
